requirements.txt
pandas
requests
yfinance
python-dotenv
matplotlib
pyyaml